+++ packages/three/index.js
@@ -1,4 +1,5 @@
 export * from "./src/ThreeRender";
 export * from "./src/Instances";
 export * from "./src/Raycast";
+export * from "./src/NavTools";